// src/contants/enums.ts
export enum EuserStatus {
    ACTIVE= 'ACTIVE',
    UNACTIVE = 'UNACTIVE',
    BANNED = 'BANNED'
}

export enum EuserRoles  {
    ADMIN = "ADMIN",
    USER = 'USER',
    EXPERT = 'EXPERT'
}

export enum EcourseStatus {
    APPROVED = "APPROVED",
    PENDING = "PENDING",
    REJECTED = "REJECTED"
}

export enum EcourseLevel {
    BEGINNER = "BEGINNER",
    INTERMEDIATE = "INTERMEDIATE",
    ADVANCE = "ADVANCE",
}

export enum ELessonType {
    VIDEO="VIDEO",
    TEXT = " TEXT"
}

export enum EOrderStatus {
    PENDING = "PENDING",
    SUCCESS = "SUCCESS",
    CANCEL = "CANCEL"
}

export enum ECouponStatus {
    ACTIVE = "ACTIVE",
    INACTIVE = "INACTIVE"
}

export enum ECouponType { 
    PERCENT = "PERCENT",
    AMOUNT = "AMOUNT"
}